Refactor Deployment CLI  (#804)
* Removed separate infrastructure module (Cleaning...)

* Update Terraform class

* Move aws funcs into aws/utils.py

* Update provider class

* Add GCP utils

* Write GCP class for the deployment

* Wrote terrascript classes to support Azurem

-- terrascript 0.9.0 does not fully support azurem provider yet :')

* Add Azure utils

* Write Azure provider for deployment

-- STILL NEED TESTINGGG

* Move CLI to domain core

* Update setup.py

* Update CLI

* Format files

* Update pyproect.toml

* Update peotry.lock

-- Add missing dependencies

* Update dependencies

* Update dependencies

* clean up CLI

* Restore correct execution script

* Decide root_dir in AWS_Serverfull class

* Fix bugs

* Install terraform plugins on the fly

* Run black tool

* Update poetry.lock

* Remove unused/outdate database_exists

* Fix bugs

* fix worker deletion bugs

* Delete worker only if already deployed

* Update Providers super class call

* Add missing Azure Resources

* Update Azure Provider class

* Update GCP Provider Class

* Update CLI

* Run black tool

+++ apps/domain/src/main/core/services/infra_service.py
@@ -230,19 +230,25 @@ def del_worker_msg(
         env = node.environments.first(id=worker_id)
         _config = Config(provider=env.provider, app=Config(name="worker", id=worker_id))
 
-        success = Provider(_config).destroy()
-        if success:
-            node.environments.set(
-                id=worker_id, state=states["destroyed"], destroyed_at=datetime.now()
+        if env.state == states["success"]:
+            worker_dir = os.path.join(
+                "/home/ubuntu/.pygrid/apps/aws/workers/", str(worker_id)
+            )
+            success = Provider(worker_dir).destroy()
+            if success:
+                node.environments.set(
+                    id=worker_id, state=states["destroyed"], destroyed_at=datetime.now()
+                )
+
+        if env.state == states["destroyed"]:
+            return DeleteWorkerResponse(
+                address=msg.reply_to,
+                status_code=200,
+                content={"msg": "Worker was deleted successfully!"},
             )
         else:
             raise Exception("Worker deletion failed")
 
-        return DeleteWorkerResponse(
-            address=msg.reply_to,
-            status_code=200,
-            content={"msg": "Worker was deleted successfully!"},
-        )
     except Exception as e:
         return DeleteWorkerResponse(
             address=msg.reply_to, status_code=False, content={"error": str(e)}
